+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Adds support to serialize the traceback for later post-mortem debugging: `ploomber build --debuglater` and `ploomber task {name} --debuglater`
 * Support for notebook post-mortem debugging: `ploomber build --debug` `ploomber task {name} --debug` ([#823](https://github.com/ploomber/ploomber/issues/823))
 * `env.yaml` allows to use existing keys in subsequent values
+* Add `start_method` to Parallel executor to set the method which should be used to start child processes ([#942](https://github.com/ploomber/ploomber/issues/942))
 * Clearing screen before updating execution status when using the `Parallel` executor
 * Fixes missing plot when generating `ploomber report` with `--backend d3` ([#946](https://github.com/ploomber/ploomber/issues/946))
 * Fixes error when using dotted paths in `grid` ([#951](https://github.com/ploomber/ploomber/issues/951))